이벤트 동기화 (오브젝트 추가 이벤트, 오브젝트 삭제 이벤트, 삭제시 Object처리)

+++ 2022_009_WinAPICore/Client/struct.h
@@ -68,6 +68,6 @@ public:
 
 	float Length()
 	{
-		return sqrt(pow(x, 2) + pow(y, 2));
+		return (float)sqrt(pow(x, 2) + pow(y, 2));
 	}
 };
